Special characters (also nodes containing `/`) (#193)
* Special characters (also nodes containing `/`)

I added tests for handling of special characters - like slashes, quotes, and unicode - in element names, paths, and values. Everything - except of one case - was working but now we know for sure.

The special case is a slash in an element name that has sub-objects: `"contains/slash": "test"` is ok, but `"contains/slash": { "test": 1 }` is not. As suggested in #169, I replaced the slash with a `tab`. Parsing now works well.

For methods that require a path as input, we could either leave it as is and remember to use a tab there, too, or we pre-process the input. 

```
{ "contains/slash": { "test": 1 } }

r->get( '/contains/slash/test' ). " never worked

r->get( |/contains\tslash/test| ). " works with this PR

r->get( '/contains\/slash/test' ). " could work if the parameter is preprocessed i.e. replace \/ with \t
```

I would go for the middle case and not worry about the parameters. 

Closes #169.

* Exclude tests from 7bit_ascii rule

* Serialize path with tab

* Fix replace + extended tests