Add support for data references (#201)
* Add support for data references

This PR adds full support for `ref to data`. You can serialize ABAP data containing fields which reference elementary, structured, or tabular data to JSON.

JSON to ABAP with data references is also possible. Prerequisite is that the references are not initial. They must point to an existing data element, structure, or table.

See the new tests for examples.

* Lint

* Move field-symbol

* Add ref initiator

* Lint

* lower case

* Renamed