Add sanity check
The PR adds a minimal sanity check. This prevents, for example, XML or plain text data to be parsed.

The check is a few miliseconds which is fine in general. It does have a significant impact on the benchmark which parses about a million JSON string.

I included a small performance improvement in the parser which should speed things up esp for larger JSON data.